contact form email

+++ src/ApplicationContainerFactory.php
@@ -10,7 +10,7 @@ class ApplicationContainerFactory
         $config = $this->getApplicationConfig();
         register_shutdown_function(function () use ($config) {
             $error = error_get_last();
-            if ($error['type'] === E_ERROR) {
+            if (isset($error['type']) && $error['type'] === E_ERROR) {
                 if (isset($config['config_cache_path']) && file_exists($config['config_cache_path'])) {
                     unlink($config['config_cache_path']);
                 }
